Add SQLAlchemy-compatible models for survey, answer, feedback
