+++ models/survey.py
@@ -1,52 +1,32 @@
+from sqlalchemy import Column, ForeignKey, Integer, String, Date
+# from sqlalchemy.ext.declarative import declarative_base
 import datetime
 
-from app import app, db
-from flask import flash, render_template, request, redirect
-from wtforms import Form, IntegerField, StringField
+from config.helper import Base
 
-class Survey:
+class Survey(Base):
     __tablename__ = "surveys"
- 
-    id = int()
-    description = str()
-    start_date = datetime.date
-    end_date = datetime.date
+
+    # Mappers
+    id_ = Column(Integer, primary_key=True)
+    description = Column(String, nullable=False)
+    start_date = Column(Date, nullable=False)
+    end_date = Column(Date, nullable=False)
 
     def __init__(self, id_, description_):
         self.id = id_
         self.description = description_
 
     def __repr__(self):
-        return "<Question: {}>".format(self.name)
+        return "<Question: {}>".format(self.description)
 
-class surveyForm(Form):
-    id = IntegerField("ID")
-    description = StringField("Description")
-
-@app.route('/addNewSurvey', methods=['GET', 'POST'])
-def addNewSurvey():
-    form = surveyForm(request.form)
-
-    if request.method == 'POST' and form.validate():
-        new_survey = db.prepare("INSERT INTO survey VALUES ($1, $2)")
-        with db.xact():
-        	new_survey(form.id.data, form.description.data)
-        
-        flash("Survey " + str(form.id.data) + " " +  \
-            form.description.data + " saved successfully!")
-        return redirect('/')
-
-    return render_template('new_survey.html', form=form)
-
-@app.route('/printSurveys', methods=['GET', 'POST'])
-def printSurveys():
-    surveys = []
-    form = surveyForm(request.form)
-
-    all_the_surveys = db.prepare("SELECT * FROM survey")
-    
-    with db.xact():
-      for survey in all_the_surveys():
-        surveys.append(survey[1])
-    
-    return render_template('surveys.html', surveys=surveys)
+    @property
+    def serialize(self):
+        start_date = self.start_date if self.start_date != None else ''
+        end_date = self.end_date if self.end_date != None else ''
+        return {
+        'id_' : self.id_,
+        'description' : self.description,
+        'start_date' : start_date,
+        'end_date' : end_date,
+        }
